Move smart action tool out of select folder
