+++ ivy-glsp-client/src/smart-action/view.tsx
